feat: simplifies data fetching model
