+++ src/components/DiagramFrame/DiagramCanvas/DiagramCanvas.tsx
@@ -26,7 +26,6 @@
 
 import React from "react"
 import {
-  Spinner,
   Heading,
   ProgressCircular,
   Status,
