+++ src/components/DiagramFrame/FramePanels/SettingsPanel/utils.tsx
@@ -23,9 +23,11 @@
  SOFTWARE.
 
  */
-
+import React from "react"
 import { SelectionInfoPacket, VisibleViewLookup } from "../../../interfaces"
-import { ILookmlModel, ILookmlModelExplore } from "@looker/sdk/lib/sdk/4.0/models"
+import { ILookmlModel, ILookmlModelExplore, IGitBranch } from "@looker/sdk/lib/sdk/4.0/models"
+import { GitBranch } from "@looker/icons"
+import { SelectOptionProps } from "@looker/components"
 import { X_INIT, Y_INIT, ZOOM_INIT, OVERRIDE_KEY_SUBTLE } from '../../../../utils/constants'
 import { internalExploreURL } from "../../../../utils/routes"
 import { ExploreDropdown } from "../types"
@@ -70,4 +72,23 @@ export function handleExploreChange (
     return OVERRIDE_KEY_SUBTLE
   }
   return undefined
+}
+
+/**
+ * prepares the Git Branch dropdown data
+ * @param gitBranch - currently selected branch obj
+ * @param gitBranches - list of available branch objs
+ */
+ export const getBranchOptions = (gitBranch: IGitBranch, gitBranches: IGitBranch[]): SelectOptionProps[] => {
+  return gitBranches?.map((branch: IGitBranch) => {
+    return gitBranch.name === branch.name ? {
+      value: branch.name, 
+      label: branch.name, 
+      icon: (<GitBranch />)
+    } : {
+      value: branch.name, 
+      label: branch.name,
+      icon: undefined
+    }
+  })
 }
